Move atomic radii to constants
