+++ src/dxtb/constants/__init__.py
@@ -4,6 +4,7 @@ Constants and default values.
 
 from . import defaults, xtb
 from .chemistry import ATOMIC_NUMBER, PSE
+from .radii import atomic_rad, cov_rad_d3, vdw_rad_d3
 from .torch import FLOAT32, FLOAT64, UINT8
 from .units import AA2AU, AU2KCAL, CAL2J, EV2AU, K2AU
 from .xtb import KA, KB, KCN, NCOORD_DEFAULT_CUTOFF, R_SHIFT
