+++ src/dxtb/data/__init__.py
@@ -1,1 +1,1 @@
-from .radii import atomic_rad, cov_rad_d3, vdw_rad_d3
+
