+++ src/dxtb/ncoord/ncoord.py
@@ -6,8 +6,7 @@ from __future__ import annotations
 import torch
 
 from .._types import Any, CountingFunction, Tensor
-from ..constants import xtb
-from ..data import cov_rad_d3
+from ..constants import xtb, cov_rad_d3
 from ..utils import real_pairs
 from .count import dexp_count, exp_count
