+++ src/dxtb/xtb/h0.py
@@ -7,8 +7,7 @@ import torch
 
 from .._types import Tensor, TensorLike
 from ..basis import IndexHelper
-from ..constants import EV2AU
-from ..data import atomic_rad
+from ..constants import EV2AU, atomic_rad
 from ..param import Param, get_elem_param, get_elem_valence, get_pair_param
 from ..utils import real_pairs, symmetrize
